requirements.txt
langchain-openai>=0.0.2
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.32.0
python-dotenv>=1.0.0 